fetch ambiguous detections that *could* belong to a tag project
uses a new upstream API entry:  project_ambiguities_for_tag_project
which returns a list of negative "ambiguity project IDs", which play
an analogous role to that of negative "ambiguity tag IDs":  they
represent a specific set of real project IDs which might be the owner
of an ambiguous tag detection.

tagme(), via motusUpdateTagDB(), now fetches all detections belonging
to any of the ambiguity project IDs, as well as those belonging to the
real project ID.  The user ends up with a database of detections that
includes all detections that are or could be of their tags.

partial fulfillment of jbrzusto/TO_DO#127